Renamed Schema.IDENTITY into Schema.BYTES
